fix: fix C++11 violation in fpm/ios.h
`fpm` targets C++11 but `fpm/ios.h` used `auto` in a lambda, which is not allowed in C++11.
This went unnoticed because the tests were compiled in C++14 mode.

To fix this, and catch these issues in the future, the tests have been switched to C++11, which required a few small changes in them.

Fixes #11