fix: fix operator<< for values at the base type's negative limit
Outputting `fpm::fixed` where the internal representation is equal to the base type's negative limit (e.g. `INT32_MIN` for `fpm::fixed_16_16`) causes an assertion / overflow because of `value.raw = -value.raw`.

To avoid rewriting the formatting logic for negative numbers, the fix is to change the formatting to use the larger `IntermediateType` from the `fpm::fixed` type, instead of `BaseType`.

Fixes #14