fix: broken conversion to float of types with maximum fraction bits
Defining fixed-point types with a maximum number of fraction bits (e.g. Q1.15 or Q1.31) and converting their `max()` or `lowest()` to float resulted in incorrect values: the sign was flipped.

This was caused by the internal fraction multiplier not being representable in the (signed) base type for those number of fraction bits. Therefore its type was extended to the larger intermediate type.

Closes #33