Only build tools when a top-level project
CPM or FetchContent are frequenly used to add dependencies to projects.
In that case, it doesn't make sense to add all these extra targets
unless they are asked for.

This is needed, for example, in the case of use in an embedded project.
When GoogleBench is included, it depends on threads, which are not
availible.